Update README.md
The example code for Grid has an error with the `remove` function. The return of `makeGrid` returns an actual function. With the code the way it is, you'd need to call `grid()` to get the remove to fire. Not `grid.remove()`. I've updated the example code to return an object parameter called `remove` that is a function that can be called how the `grid.remove()` pattern is attempting at the end.